fix: move PaginatedReviewBase component to components dir
