+++ src/pages/review/dalle.tsx
@@ -1,5 +1,5 @@
 import { ImageOrigin } from "@/lib/enums/ImageOrigin";
-import { PaginatedReviewBase } from "./PaginatedReviewBase";
+import { PaginatedReviewBase } from "../../components/PaginatedReviewBase";
 
 export default function ReviewDalleTransfers() {
     return <PaginatedReviewBase origin={ImageOrigin.Dalle} />;
